split the connections processes into two classes
